--- src/lmnr/openllmetry_sdk/tracing/__init__.py
@@ -1,1 +1,0 @@
-from lmnr.openllmetry_sdk.tracing.context_manager import get_tracer
