+++ src/lmnr/opentelemetry_lib/opentelemetry/instrumentation/google_genai/utils.py
@@ -34,7 +34,7 @@ def dont_throw(func):
             return func(*args, **kwargs)
         except Exception as e:
             logger.debug(
-                "OpenLLMetry failed to trace in %s, error: %s",
+                "Laminar failed to trace in %s, error: %s",
                 func.__name__,
                 traceback.format_exc(),
             )
