+++ src/lmnr/opentelemetry_lib/tracing/__init__.py
@@ -1,0 +1,1 @@
+from lmnr.opentelemetry_lib.tracing.context_manager import get_tracer
