+++ src/lmnr/opentelemetry_lib/utils/json_encoder.py
