+++ src/lmnr/sdk/evaluations.py
@@ -5,8 +5,8 @@ import dotenv
 from tqdm import tqdm
 from typing import Any, Awaitable, Optional, Set, Union
 
-from lmnr.openllmetry_sdk.instruments import Instruments
-from lmnr.openllmetry_sdk.tracing.attributes import SPAN_TYPE
+from lmnr.opentelemetry_lib.instruments import Instruments
+from lmnr.opentelemetry_lib.tracing.attributes import SPAN_TYPE
 
 from lmnr.sdk.client.asynchronous.async_client import AsyncLaminarClient
 from lmnr.sdk.client.synchronous.sync_client import LaminarClient
@@ -176,7 +176,6 @@ class Evaluation:
 
         base_url = base_url or from_env("LMNR_BASE_URL") or "https://api.lmnr.ai"
 
-        self.is_finished = False
         self.reporter = EvaluationReporter(base_url)
         if isinstance(data, list):
             self.data = [
@@ -225,8 +224,6 @@ class Evaluation:
         )
 
     async def run(self) -> Awaitable[None]:
-        if self.is_finished:
-            raise Exception("Evaluation is already finished")
         return await self._run()
 
     async def _run(self) -> None:
@@ -253,13 +250,11 @@ class Evaluation:
                 self._logger.debug("All upload tasks completed")
         except Exception as e:
             self.reporter.stopWithError(e)
-            self.is_finished = True
             await self._shutdown()
             return
 
         average_scores = get_average_scores(result_datapoints)
         self.reporter.stop(average_scores, evaluation.projectId, evaluation.id)
-        self.is_finished = True
         await self._shutdown()
 
     async def _shutdown(self):
